requirements.txt
PyQt6>=6.6.0
pdfplumber>=0.10.0
python-pptx>=0.6.23
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.32
langchain-openai>=0.0.8
langchain-anthropic>=0.1.11
chromadb>=0.4.0
openai>=1.14.0
anthropic>=0.18.0
PyInstaller>=6.4.0
keyring>=25.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
pyjwt>=2.8.0
